GamePlusJames ActionRPGTutorial completed to ep40
Followed a tutorial series , the series used an older version of Unity,
which was versoin 5.2.0.f3. the Unity which was used to follow the
series was verison 5.5.0f3.  there are some differences between
verisons. Also the maps that are made are made tried to replicate the
maps used in the tutorial, but they are not the same.